refactor: remove unnecessary int64 conversions and improve code quality
- Remove unnecessary int64() type conversions where types already match
- Replace for loops with copy() for slice copying (staticcheck S1001)
- Fix hookid parsing to use ParseInt instead of Atoi for int64 types
- Simplify pointer assignments by directly using source pointers
- Fix project member UserID to return pointer as expected by API

Changes:
- Use copy() instead of manual loops in approvalrules.go
- Remove redundant int64() casts throughout client and controller code
- Update strconv.Atoi to strconv.ParseInt for int64 variables
- Simplify group and project observation field assignments

All tests passing (55 packages).
Build successful.

+++ pkg/cluster/clients/groups/zz_group_test.go
@@ -104,7 +104,7 @@ func TestGenerateObservation(t *testing.T) {
 		"Full": {
 			args: args{
 				p: &gitlab.Group{
-					ID:         int64(id),
+					ID:         id,
 					WebURL:     webURL,
 					FullName:   fullName,
 					FullPath:   fullPath,
