+++ pkg/cluster/controller/projects/badges/zz_controller.go
@@ -175,7 +175,7 @@ func (e *external) Create(ctx context.Context, mg resource.Managed) (managed.Ext
 	// if ID is already set, check if it does exist, else create a new one
 	if cr.Spec.ForProvider.ID != nil {
 		id := *cr.Spec.ForProvider.ID
-		badge, res, err := e.client.GetProjectBadge(*cr.Spec.ForProvider.ProjectID, int64(id))
+		badge, res, err := e.client.GetProjectBadge(*cr.Spec.ForProvider.ProjectID, id)
 		if err != nil || clients.IsResponseNotFound(res) {
 			return managed.ExternalCreation{}, errors.Wrap(err, errWrongIDSet)
 		}
